Removed class `ExtendedPayments`
