--- src/Features/OnboardingTasks/Tasks/ExtendedPayments.php
@@ -1,90 +1,0 @@
-<?php
-
-namespace Automattic\WooCommerce\Admin\Features\OnboardingTasks\Tasks;
-
-use Automattic\WooCommerce\Admin\Features\Features;
-use Automattic\WooCommerce\Admin\Features\OnboardingTasks\Tasks\WooCommercePayments;
-use Automattic\WooCommerce\Admin\Features\OnboardingTasks\Task;
-
-/**
- * Payments Task
- */
-class ExtendedPayments extends Task {
-	/**
-	 * ID.
-	 *
-	 * @return string
-	 */
-	public function get_id() {
-		return 'payments';
-	}
-
-	/**
-	 * Title.
-	 *
-	 * @return string
-	 */
-	public function get_title() {
-		return __( 'Set up additional payment providers', 'woocommerce-admin' );
-	}
-
-	/**
-	 * Content.
-	 *
-	 * @return string
-	 */
-	public function get_content() {
-		return __(
-			'Choose payment providers and enable payment methods at checkout.',
-			'woocommerce-admin'
-		);
-	}
-
-	/**
-	 * Time.
-	 *
-	 * @return string
-	 */
-	public function get_time() {
-		return __( '2 minutes', 'woocommerce-admin' );
-	}
-
-	/**
-	 * Task completion.
-	 *
-	 * @return bool
-	 */
-	public function is_complete() {
-		return self::has_gateways();
-	}
-
-	/**
-	 * Task visibility.
-	 *
-	 * @return bool
-	 */
-	public function can_view() {
-		return Features::is_enabled( 'payment-gateway-suggestions' ) &&
-			WooCommercePayments::is_requested() &&
-			WooCommercePayments::is_installed() &&
-			WooCommercePayments::is_supported() &&
-			( $this->get_parent_id() !== 'extended_two_column' || ! WooCommercePayments::is_connected() );
-	}
-
-	/**
-	 * Check if the store has any enabled gateways.
-	 *
-	 * @return bool
-	 */
-	public static function has_gateways() {
-		$gateways         = WC()->payment_gateways->get_available_payment_gateways();
-		$enabled_gateways = array_filter(
-			$gateways,
-			function( $gateway ) {
-				return 'yes' === $gateway->enabled && 'woocommerce_payments' !== $gateway->id;
-			}
-		);
-
-		return ! empty( $enabled_gateways );
-	}
-}
